split types into smaller files
https://coveord.atlassian.net/browse/KIT-2600
